requirements.txt
google-genai>=1.0.0
openai>=1.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
